Added proper checkbox and radio button images
Added a few images to help with the checkbox and radio button design.  I
changed the HTML tags on some of the elements. I removed the H2 tag and
used the legend where appropriate. I need to go back and style the CSS
and check for unused classes.  Also need to change the arrow style of
the chevron in the State select box.